code/auto_expansion_sysml/requirements.txt
# Core dependencies
pymilvus>=2.3.0
pymongo>=4.5.0
neo4j>=5.13.0
openai>=1.0.0
langchain>=0.0.300
langchain-openai>=0.0.5
python-dotenv>=1.0.0
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.4.2
numpy>=1.24.3
pandas>=2.0.3
tqdm>=4.66.1
requests>=2.31.0
python-multipart>=0.0.6 